backend/requirements.txt
django
djangorestframework
python-dotenv
mysqlclient
django-cors-headers
sib-api-v3-sdk==7.6.0